+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
  * Add options `inline_factories` and `inline_class_loader` to `PhpDumper::dump()`
  * Deprecate `PhpDumper` options `inline_factories_parameter` and `inline_class_loader_parameter`
  * Add `RemoveBuildParametersPass`, which removes parameters starting with a dot during compilation
+ * Add support for nesting autowiring-related attributes into `#[Autowire(...)]`
 
 6.2
 ---